Fixed fucked relationships part 1 of 7

+++ db/schema.rb
@@ -44,7 +44,7 @@ ActiveRecord::Schema.define(version: 2019_06_14_184815) do
     t.string "password_digest"
     t.string "email"
     t.integer "phone"
-    t.integer "wallet", default: 0
+    t.float "wallet", default: 0.0
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
